+++ pymc3/tests/test_tree_nodes.py
@@ -27,20 +27,20 @@ def test_bad_split_nodes_creation():
 
 
 def test_good_leaf_node_creation():
-    leaf_node = LeafNode(index=0, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int64'))
+    leaf_node = LeafNode(index=0, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int32'))
     assert leaf_node.index == 0
-    assert np.array_equal(leaf_node.idx_data_points, np.array([1, 2, 3], dtype='int64'))
+    assert np.array_equal(leaf_node.idx_data_points, np.array([1, 2, 3], dtype='int32'))
     assert leaf_node.value == 22.2
 
-    leaf_node = LeafNode(index=0, value=22.2, idx_data_points=np.array([1, 6, 7], dtype='int64'))
+    leaf_node = LeafNode(index=0, value=22.2, idx_data_points=np.array([1, 6, 7], dtype='int32'))
     assert leaf_node.index == 0
-    assert np.array_equal(leaf_node.idx_data_points, np.array([1, 6, 7], dtype='int64'))
+    assert np.array_equal(leaf_node.idx_data_points, np.array([1, 6, 7], dtype='int32'))
     assert leaf_node.value == 22.2
 
 
 def test_bad_leaf_nodes_creation():
     with pytest.raises(TreeNodeError) as err:
-        LeafNode(index=-1, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int64'))
+        LeafNode(index=-1, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int32'))
     assert str(err.value) == 'Node index must be a non-negative int'
 
     with pytest.raises(TreeNodeError) as err:
@@ -48,11 +48,11 @@ def test_bad_leaf_nodes_creation():
     assert str(err.value) == 'Index of data points must be a numpy.ndarray of integers'
 
     with pytest.raises(TreeNodeError) as err:
-        LeafNode(index=0, value=22.2, idx_data_points=np.array([], dtype='int64'))
+        LeafNode(index=0, value=22.2, idx_data_points=np.array([], dtype='int32'))
     assert str(err.value) == 'Index of data points can not be empty'
 
     with pytest.raises(TreeNodeError) as err:
-        LeafNode(index=0, value='2', idx_data_points=np.array([1, 2, 3], dtype='int64'))
+        LeafNode(index=0, value='2', idx_data_points=np.array([1, 2, 3], dtype='int32'))
     assert str(err.value) == 'Leaf node value type must be float'
 
 
@@ -71,7 +71,7 @@ def test_correct_get_idx_parent_node():
     node1 = SplitNode(index=13, idx_split_variable=2, split_value=3.0)
     assert node1.get_idx_parent_node() == 6
 
-    node2 = LeafNode(index=4, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int64'))
+    node2 = LeafNode(index=4, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int32'))
     assert node2.get_idx_parent_node() == 1
 
 
@@ -79,7 +79,7 @@ def test_correct_get_idx_left_child():
     node1 = SplitNode(index=3, idx_split_variable=2, split_value=3.0)
     assert node1.get_idx_left_child() == 7
 
-    node2 = LeafNode(index=4, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int64'))
+    node2 = LeafNode(index=4, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int32'))
     assert node2.get_idx_left_child() == 9
 
 
@@ -87,7 +87,7 @@ def test_correct_get_idx_right_child():
     node1 = SplitNode(index=3, idx_split_variable=2, split_value=3.0)
     assert node1.get_idx_right_child() == 8
 
-    node2 = LeafNode(index=4, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int64'))
+    node2 = LeafNode(index=4, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int32'))
     assert node2.get_idx_right_child() == 10
 
 
@@ -95,7 +95,7 @@ def test_correct_is_left_child():
     node1 = SplitNode(index=3, idx_split_variable=2, split_value=3.0)
     assert node1.is_left_child() is True
 
-    node2 = LeafNode(index=4, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int64'))
+    node2 = LeafNode(index=4, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int32'))
     assert node2.is_left_child() is False
 
 
@@ -103,7 +103,7 @@ def test_correct_get_idx_sibling():
     node1 = SplitNode(index=3, idx_split_variable=2, split_value=3.0)
     assert node1.get_idx_sibling() == 4
 
-    node2 = LeafNode(index=11, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int64'))
+    node2 = LeafNode(index=11, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int32'))
     assert node2.get_idx_sibling() == 12
 
 
@@ -118,11 +118,11 @@ def test_correct_nodes_eq():
     assert (split_node1 == split_node4) is False
     assert (split_node1 == split_node5) is False
 
-    leaf_node1 = LeafNode(index=1, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int64'))
-    leaf_node2 = LeafNode(index=1, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int64'))
-    leaf_node3 = LeafNode(index=2, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int64'))
-    leaf_node4 = LeafNode(index=1, value=55.5, idx_data_points=np.array([1, 2, 3], dtype='int64'))
-    leaf_node5 = LeafNode(index=1, value=55.5, idx_data_points=np.array([1, 2], dtype='int64'))
+    leaf_node1 = LeafNode(index=1, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int32'))
+    leaf_node2 = LeafNode(index=1, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int32'))
+    leaf_node3 = LeafNode(index=2, value=22.2, idx_data_points=np.array([1, 2, 3], dtype='int32'))
+    leaf_node4 = LeafNode(index=1, value=55.5, idx_data_points=np.array([1, 2, 3], dtype='int32'))
+    leaf_node5 = LeafNode(index=1, value=55.5, idx_data_points=np.array([1, 2], dtype='int32'))
     assert (leaf_node1 == leaf_node2) is True
     assert (leaf_node1 == leaf_node3) is False
     assert (leaf_node1 == leaf_node4) is False
